@@ -16,32 +16,32 @@
     "test:ci": "yarn run lint && yarn test"
   },
   "devDependencies": {
-    "babel-core": "~6.22.1",
+    "babel-core": "~6.23.1",
     "babel-eslint": "~7.1.1",
-    "babel-loader": "~6.2.9",
+    "babel-loader": "~6.3.0",
     "babel-plugin-__coverage__": "~11.0.0",
     "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-preset-airbnb": "~2.1.1",
-    "babel-preset-es2015": "~6.18.0",
-    "babel-preset-react": "~6.16.0",
-    "babel-preset-stage-2": "~6.18.0",
+    "babel-preset-airbnb": "~2.2.3",
+    "babel-preset-es2015": "~6.22.0",
+    "babel-preset-react": "~6.22.0",
+    "babel-preset-stage-2": "~6.22.0",
     "chai": "~3.5.0",
     "chai-as-promised": "~6.0.0",
     "chai-enzyme": "~0.6.1",
-    "enzyme": "~2.6.0",
-    "eslint": "~3.8.1",
+    "enzyme": "~2.7.1",
+    "eslint": "~3.15.0",
     "eslint-loader": "~1.6.1",
     "eslint-plugin-import": "~2.2.0",
     "eslint-plugin-jsx-a11y": "~4.0.0",
-    "eslint-plugin-react": "~6.9.0",
+    "eslint-plugin-react": "~6.10.0",
     "expose-loader": "~0.7.1",
     "faker": "~3.1.0",
     "jasmine-core": "~2.5.2",
     "json-loader": "~0.5.4",
-    "karma": "~1.3.0",
+    "karma": "~1.4.1",
     "karma-chai": "~0.1.0",
     "karma-coverage": "~1.1.1",
-    "karma-jasmine": "~1.0.2",
+    "karma-jasmine": "~1.1.0",
     "karma-phantomjs-launcher": "~1.0.2",
     "karma-sinon-chai": "~1.2.4",
     "karma-sourcemap-loader": "~0.3.7",
@@ -52,9 +52,9 @@
     "sinon": "~1.17.6",
     "sinon-chai": "~2.8.0",
     "webpack": "~2.2.1",
-    "webpack-bundle-analyzer": "~2.2.1",
+    "webpack-bundle-analyzer": "~2.3.0",
     "webpack-config-utils": "~2.3.0",
-    "webpack-validator": "~2.2.12",
+    "webpack-validator": "~2.3.0",
     "yaml-loader": "~0.4.0"
   },
   "dependencies": {
@@ -66,7 +66,7 @@
     "moment": "~2.17.1",
     "react": "~15.4.2",
     "react-addons-update": "~15.4.2",
-    "react-apollo": "~0.10.1",
+    "react-apollo": "~0.11.1",
     "react-dom": "~15.4.2",
     "react-i18nify": "~1.8.4",
     "uuid": "~3.0.1"
