@@ -26,7 +26,7 @@ class CommentOrderSelector extends Component {
     return (
       <div className="order-by__dropdown order-by__dropdown--right">
         <span className="order-by__text">{ I18n.t("components.comment_order_selector.title") }</span>
-        <ul className="dropdown menu" data-dropdown-menu>
+        <ul className="dropdown menu" data-dropdown-menu data-close-on-click-inside="false">
           <li>
             <a>{ I18n.t(`components.comment_order_selector.order.${orderBy}`) }</a>
             <ul className="menu">
