@@ -15,10 +15,10 @@ import generateUserData     from '../support/generate_user_data';
 import resolveGraphQLQuery  from '../support/resolve_graphql_query';
 
 describe('<Comments />', () => {
-  let comments = [];
+  let commentable = {};
   let session = null;
   const commentableId = "1";
-  const commentableType = "Decidim::ParticipatoryProcess";
+  const commentableType = "Decidim::DummyResource";
   const orderBy = "older";
   const reorderComments = () => {};
 
@@ -28,11 +28,17 @@ describe('<Comments />', () => {
     }
   `;
 
-  const addCommentFragment = gql`
-    fragment AddCommentForm on User {
+  const addCommentFormSessionFragment = gql`
+    fragment AddCommentFormSession on Session {
       verifiedUserGroups {
         id
       }
+    }
+  `;
+
+  const addCommentFormCommentableFragment = gql`
+    fragment AddCommentFormCommentable on Commentable {
+      id
     }
   `;
 
@@ -46,7 +52,8 @@ describe('<Comments />', () => {
 
   stubComponent(AddCommentForm, {
     fragments: {
-      user: addCommentFragment
+      session: addCommentFormSessionFragment,
+      commentable: addCommentFormCommentableFragment
     }
   });
 
@@ -57,7 +64,8 @@ describe('<Comments />', () => {
     const query = gql`
       ${commentsQuery}
       ${commentThreadFragment}
-      ${addCommentFragment}
+      ${addCommentFormSessionFragment}
+      ${addCommentFormCommentableFragment}
     `;
 
     const result = resolveGraphQLQuery(query, {
@@ -66,7 +74,12 @@ describe('<Comments />', () => {
         session: {
           user: userData
         },
-        comments: commentsData
+        commentable: {
+          acceptsNewComments: true,
+          commentsHaveAlignment: true,
+          commentsHaveVotes: true,
+          comments: commentsData
+        }
       },
       variables: {
         orderBy,
@@ -76,82 +89,82 @@ describe('<Comments />', () => {
     });
 
     session = result.session;
-    comments = result.comments;
+    commentable = result.commentable;
   });
 
-  it("should render loading-comments calss and the respective loading text", () => {
-    const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} loading />);
+  it("renders loading-comments class and the respective loading text", () => {
+    const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} loading />);
     expect(wrapper.find('.loading-comments')).to.be.present();
     expect(wrapper.find('h2')).to.have.text("Loading comments ...");
   });
 
-  it("should render a div of id comments", () => {
-    const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
+  it("renders a div of id comments", () => {
+    const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
     expect(wrapper.find('#comments')).to.be.present();
   });
 
-  describe("should render a CommentThread component for each comment", () => {
+  describe("renders a CommentThread component for each comment", () => {
     it("and pass filter comment data as a prop to it", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
-      expect(wrapper).to.have.exactly(comments.length).descendants(CommentThread);
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
+      expect(wrapper).to.have.exactly(commentable.comments.length).descendants(CommentThread);
       wrapper.find(CommentThread).forEach((node, idx) => {
-        expect(node).to.have.prop("comment").deep.equal(filter(commentThreadFragment, comments[idx]));
+        expect(node).to.have.prop("comment").deep.equal(filter(commentThreadFragment, commentable.comments[idx]));
       });
     });
 
     it("and pass the session as a prop to it", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
-      expect(wrapper).to.have.exactly(comments.length).descendants(CommentThread);
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
+      expect(wrapper).to.have.exactly(commentable.comments.length).descendants(CommentThread);
       wrapper.find(CommentThread).forEach((node) => {
         expect(node).to.have.prop("session").deep.equal(session);
       });
     });
 
-    it("and pass the option votable as a prop to it", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{ votable: true }} reorderComments={reorderComments} orderBy={orderBy} />);
-      expect(wrapper).to.have.exactly(comments.length).descendants(CommentThread);
+    it("and pass the commentable 'commentsHaveVotes' property as a prop to it", () => {
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
+      expect(wrapper).to.have.exactly(commentable.comments.length).descendants(CommentThread);
       wrapper.find(CommentThread).forEach((node) => {
         expect(node).to.have.prop("votable").equal(true);
       });
     });
   });
 
-  it("should render comments count", () => {
-    const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
-    const rex = new RegExp(`${comments.length} comments`);
+  it("renders comments count", () => {
+    const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
+    const rex = new RegExp(`${commentable.comments.length} comments`);
     expect(wrapper.find('h2.section-heading')).to.have.text().match(rex);
   });
 
-  it("should render a AddCommentForm component and pass 'options.arguable' as a prop", () => {
-    const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{ arguable: true }} reorderComments={reorderComments} orderBy={orderBy} />);
+  it("renders a AddCommentForm component and pass the commentable 'commentsHaveAlignment' as a prop", () => {
+    const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
     expect(wrapper).to.have.exactly(1).descendants(AddCommentForm);
     expect(wrapper.find(AddCommentForm)).to.have.prop('arguable').equal(true);
   });
 
-  describe("if session is not present", () => {
+  describe("when the commentable cannot accept new comments", () => {
     beforeEach(() => {
-      session = null;
+      commentable.acceptsNewComments = false;
     });
 
-    it("should not render a AddCommentForm component", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
+    it("doesn't render an AddCommentForm component", () => {
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
       expect(wrapper.find(AddCommentForm)).not.to.be.present();
+    });
+
+    it("renders a callout message to inform the user that comments are blocked", () => {
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
+      expect(wrapper.find('.callout.warning')).to.include.text("disabled");
     });
   });
 
-  describe("should render a CommentOrderSelector component", () => {
-    it("should render a CommentOrderSelector component", () => {
-        const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
-        expect(wrapper.find(CommentOrderSelector)).to.be.present();
-      });
-
+  describe("renders a CommentOrderSelector component", () => {
     it("and pass the reorderComments as a prop to it", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
       expect(wrapper.find(CommentOrderSelector)).to.have.prop('reorderComments').deep.equal(reorderComments);
     });
 
     it("and pass the orderBy as a prop to it", () => {
-      const wrapper = shallow(<Comments comments={comments} commentableId={commentableId} commentableType={commentableType} session={session} options={{}} reorderComments={reorderComments} orderBy={orderBy} />);
+      const wrapper = shallow(<Comments commentable={commentable} session={session} reorderComments={reorderComments} orderBy={orderBy} />);
       expect(wrapper.find(CommentOrderSelector)).to.have.prop('defaultOrderBy').equal('older');
     });
   });
