@@ -46,7 +46,7 @@ UpVoteButton.propTypes = {
 
 const UpVoteButtonWithMutation = graphql(gql`
   ${upVoteMutation}
-  ${commentFragment}  
+  ${commentFragment}
   ${commentDataFragment}
   ${upVoteFragment}
   ${downVoteFragment}
@@ -71,20 +71,23 @@ const UpVoteButtonWithMutation = graphql(gql`
       updateQueries: {
         GetComments: (prev, { mutationResult: { data } }) => {
           const commentReducer = (comment) => {
-            const replies = comment.replies || [];
+            const replies = comment.comments || [];
 
             if (comment.id === ownProps.comment.id) {
               return data.comment.upVote;
             }
             return {
               ...comment,
-              replies: replies.map(commentReducer)
+              comments: replies.map(commentReducer)
             };
           };
 
           return {
             ...prev,
-            comments: prev.comments.map(commentReducer)
+            commentable: {
+              ...prev.commentable,
+              comments: prev.commentable.comments.map(commentReducer)
+            }
           }
         }
       }
