@@ -1,6 +1,6 @@
 import { random, name, date, image } from 'faker/locale/en';
 
-/** 
+/**
  * Generate random comment data to emulate a database real content
  * @param {number} num - The number of comments to generate random data
  * @returns {Object[]} - An array of objects representing comments data
@@ -17,9 +17,9 @@ const generateCommentsData = (num = 1) => {
         name: name.findName(),
         avatarUrl: image.imageUrl()
       },
-      hasReplies: false,
-      replies: [],
-      canHaveReplies: true,
+      hasComments: false,
+      comments: [],
+      acceptsNewComments: true,
       alignment: 0,
       upVotes: random.number(),
       upVoted: false,
