@@ -17,12 +17,7 @@ module Decidim
       end
 
       def current_settings
-        @current_settings ||= begin
-          active_step = current_participatory_process.active_step
-          return nil unless active_step
-
-          current_feature.step_settings.fetch(active_step.id.to_s)
-        end
+        @current_settings ||= current_feature.active_step_settings
       end
     end
   end
