@@ -11,7 +11,13 @@ module Decidim
 
     authorize_resource :public_pages, class: false
     delegate :page, to: :page_finder
-    helper_method :page, :promoted_participatory_processes, :participatory_processes, :users
+    helper_method :page, :promoted_participatory_processes, :highlighted_participatory_processes, :participatory_processes, :users
+
+    def index
+      @pages = current_organization.static_pages.all.to_a.sort do |a, b|
+        a.title[I18n.locale.to_s] <=> b.title[I18n.locale.to_s]
+      end
+    end
 
     def page_finder
       @page_finder ||= Decidim::PageFinder.new(params[:id], current_organization)
@@ -27,7 +33,11 @@ module Decidim
     end
 
     def promoted_participatory_processes
-      @promoted_processes ||= OrganizationParticipatoryProcesses.new(current_organization) | PublicParticipatoryProcesses.new | PromotedParticipatoryProcesses.new
+      @promoted_processes ||= participatory_processes | PromotedParticipatoryProcesses.new
+    end
+
+    def highlighted_participatory_processes
+      @promoted_processes ||= OrganizationParticipatoryProcesses.new(current_organization) | HighlightedParticipatoryProcesses.new
     end
   end
 end
