@@ -4,7 +4,7 @@ module Decidim
   # public view.
   class PublicParticipatoryProcesses < Rectify::Query
     def query
-      Decidim::ParticipatoryProcess.includes(:active_step).order("end_date ASC")
+      Decidim::ParticipatoryProcess.order("promoted DESC").includes(:active_step).order("decidim_participatory_process_steps.end_date ASC")
     end
   end
 end
