@@ -41,8 +41,8 @@ FactoryGirl.define do
     welcome_text { Decidim::Faker::Localized.wrapped("<p>", "</p>") { Decidim::Faker::Localized.sentence(2) } }
     homepage_image { test_file("city.jpeg", "image/jpeg") }
     favicon { test_file("icon.png", "image/png") }
-    default_locale I18n.default_locale
-    available_locales Decidim.available_locales
+    default_locale { I18n.default_locale }
+    available_locales { Decidim.available_locales }
     official_img_header { test_file("avatar.jpg", "image/jpeg") }
     official_img_footer { test_file("avatar.jpg", "image/jpeg") }
     official_url { Faker::Internet.url }
@@ -225,6 +225,14 @@ FactoryGirl.define do
     name { generate(:slug) }
     to { build(:dummy_resource) }
     from { build(:dummy_resource, feature: to.feature) }
+  end
+
+  factory :newsletter, class: Decidim::Newsletter do
+    author { build(:user, :confirmed, organization: organization) }
+    organization
+
+    subject { Decidim::Faker::Localized.sentence(3) }
+    body { Decidim::Faker::Localized.wrapped("<p>", "</p>") { Decidim::Faker::Localized.sentence(4) } }
   end
 end
 
