@@ -9,30 +9,37 @@ RSpec.shared_examples "with participatory processes" do
       it "orders them by end_date" do
         unpublished = create(
           :participatory_process,
+          :with_steps,
           :unpublished,
           organization: organization
         )
 
         last = create(
           :participatory_process,
+          :with_steps,
           :published,
           organization: organization,
           end_date: nil
         )
+        last.active_step.update_attribute(:end_date, nil)
 
         first = create(
           :participatory_process,
+          :with_steps,
           :published,
           organization: organization,
           end_date: Time.current.advance(days: 10)
         )
+        first.active_step.update_attribute(:end_date, Time.current.advance(days: 2))
 
         second = create(
           :participatory_process,
+          :with_steps,
           :published,
           organization: organization,
           end_date: Time.current.advance(days: 20)
         )
+        second.active_step.update_attribute(:end_date, Time.current.advance(days: 4))
 
         expect(controller.helpers.participatory_processes.count).to eq(3)
         expect(controller.helpers.participatory_processes).not_to include(unpublished)
@@ -50,9 +57,10 @@ RSpec.shared_examples "with promoted participatory processes" do
   end
   describe "helper methods" do
     describe "promoted_participatory_processes" do
-      it "orders them by end_date" do
+      it "orders them by active_step end_date" do
         unpublished = create(
           :participatory_process,
+          :with_steps,
           :unpublished,
           :promoted,
           organization: organization
@@ -60,33 +68,39 @@ RSpec.shared_examples "with promoted participatory processes" do
 
         unpromoted = create(
           :participatory_process,
+          :with_steps,
           :unpublished,
           organization: organization
         )
 
         last = create(
           :participatory_process,
+          :with_steps,
           :published,
           :promoted,
-          organization: organization,
-          end_date: nil
+          organization: organization
         )
+        last.active_step.update_attribute(:end_date, nil)
 
         first = create(
           :participatory_process,
+          :with_steps,
           :published,
           :promoted,
           organization: organization,
           end_date: Time.current.advance(days: 10)
         )
+        first.active_step.update_attribute(:end_date, Time.current.advance(days: 2))
 
         second = create(
           :participatory_process,
+          :with_steps,
           :published,
           :promoted,
           organization: organization,
-          end_date: Time.current.advance(days: 20)
+          end_date: Time.current.advance(days: 8)
         )
+        second.active_step.update_attribute(:end_date, Time.current.advance(days: 4))
 
         expect(controller.helpers.promoted_participatory_processes.count).to eq(3)
         expect(controller.helpers.promoted_participatory_processes).not_to include(unpublished)
