@@ -17,11 +17,11 @@ Decidim.register_feature(:results) do |feature|
   end
 
   feature.settings(:global) do |settings|
-    settings.attribute :comments_always_enabled, type: :boolean, default: true
+    settings.attribute :comments_enabled, type: :boolean, default: true
   end
 
   feature.settings(:step) do |settings|
-    settings.attribute :comments_enabled, type: :boolean, default: true
+    settings.attribute :comments_blocked, type: :boolean, default: false
   end
 
   feature.seeds do
